requirements.txt
aisuite
albumentationsx
cleanlab
matplotlib
numpy
pydantic
pandas
opencv-python
timm
torch